More work on the initramfs
This finishes an initial stab at the initramfs. We have the basic
process of creating the initramfs working, and we can now boot into a
shell.

Next step is to clean up the code to take a dynamic list of files to
include, automatically infer the correct file modes, and remove
the path duplication.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>