Update display as we type
Because we disable `echo` mode when we start the shell, we need to
manually update the display as we type. This adds in a call to write
out each character as it is typed.

We also update the `Buffer` to take a `Write` object, so that we can
write to the terminal directly, without having to worry about any line
buffering that might be going on.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>