Add support for pipelining to the shell
This adds support for pipelining to the shell. Pipelining
allows us to chain together multiple commands, with the
output of one command being passed as the input to the
next command.

This works by creating a process group and adding
all the commands to the process group. We then
wait on all of them to finish, and then return the
exit status of the last command in the pipeline.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>