Don't probe device when specifying a type manually
In `mount`, we always try to probe the device to determine the filesystem
but this isn't necessary if the user has specified the filesystem type
in the command line. This change makes it so that we only probe the
device if the filesystem type is not specified, or if the type is `auto`.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>