Make sure that the tty is opened in R/W mode
When we open a TTY in getty, we allow the user to
pass in `-`, which means that they've already
opened it for us (i.e. use the existing stdin).

This adds in a check to make sure that the TTY is
opened in R/W mode, as we need to be able to write
to it as well as read from it.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>