Create `common` crate for shared code
This commit creates a new crate called `common` which will contain
common utils that are shared between the different parts of the
operating system.

To start off, we have an `io` module that brings in the
`IOTriple` struct from the shell (which allows getting unbuffered
handles to the std streams), and an `obs` module that brings in
the logic for setting up a logger.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>