Add `login` daemon
This adds a simple `login` daemon that will read the username and
password from the terminal and check it against the `passwd` and
`shadow` files. If the user is authenticated, it will then exec the
user's shell.

There's a bit more to do here - we need to set the uid/gid
appropriately, and do the chdir to the user's home directory, but this
allows us to log in which is pretty neat.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>