Add `Size` trait to `bytestruct` and `bytestruct-derive`
For padding, we want to know the size of a type so that
we can calculate the padding required to align it to
a certain boundary.

This commit adds a `Size` trait to `bytestruct` and
`bytestruct-derive` that provides a `size` method that
returns the size of the type in bytes.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>