Support dropping permissions in qinit
This adds the ability for services to set a user
and group that they should run as. qinit will
set the user and group before execing the service.

Optionally, we can also create a new user and group
for the service to run as, using the new create
methods in the auth crate.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>