Overhaul service starting in qinit
Previously, when we wanted to start a service, we would
topologically sort the dependencies and then start one at a time.
This works when we can predict the start times of services,
but this is changing.

In particular, I'm introducing the idea of "Start modes", allowing
services to mark themselves as _running_, after having been started
(analogous to ready checks in K8s). Services that depend on
a starting service will not have that dependency satisfied (and
thus will not be started) until its dependencies are marked running
not just started. Because we can't predict how long that will
take to happen, blocking a bunch of starts doesn't work anymore.

To make this better, we instead introduce "service watchers".
Service watchers will represent services that are waiting
to start. Every time a service is marked as running,
we notify all the service watchers and then start any that
are now ready.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>