Support partial strings in reading string tables
As an optimisation, it seems like indices into
string sections can be to the middle of strings
to obtain substrings and avoid duplication.

The previous implementation assumed that all
strings started after the preceeding \0, so we
weren't resolving these partial strings properly.

This adds in extra logic to extract substrings when
the offset is in the middle of a string.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>