Implement framing in Bus
Previously, bus would buffer the entire contents of a socket
connection into a single message, so if you wanted
to create multiple messages then you would need
to open a publish channel for each one.

This implements framing so that we can use the same
channel for multiple messages. Each message to be sent
is prepended with a u16 containing the length of the
message, and we augment the subscription api to read
that and populate the buffer correctly.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>