Properly type Netlink groups
Previously we were using raw u32's. This adds in an associated type to the
Netlink trait that allows the specific types to define their own Group enum
so that groups can be properly typed.

We then define those groups for both the existing Netlink specialisations -
Uevents and RTNetlink. This will allow us to subscribe to link updates
etc for the upcoming netd implementation.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>