Make NetlinkSocket flow aware
Multiple flows are multiplexed onto the one netlink socket -
one message might be for the multicast groups, another might be
responding to a request etc. Previously, we assumed that any
new message is responding to the immediately previous request
however with netd this is no longer the case.

To solve this, we've rejigged how reading from a NetlinkSocket works.
Instead of each consumer reading directly, we instead have an mpsc
channel that gets sent back after each request. We then have a background
process that reads messages from the socket and dispatches
them to the correct channel depending on their sequence number.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>